More BMS accuracy fixes.

+++ src/SongTiming.h
@@ -5,31 +5,31 @@ inline float spb(float bpm) { return 60 / bpm; } // Return seconds per beat.
 inline float bps(float bpm) { return bpm / 60; } // Return beats per second.
 
 /* Assume these are sorted. */
-// Return the (vector's index+1) Beat resides in.
-uint32 SectionIndex(const TimingData &Timing, float Beat);
+// Return the vector's index Beat resides in.
+int SectionIndex(const TimingData &Timing, double Beat);
 
 /*
 	Find the value of the interval function defined by Timing at the value Beat
 */
-double SectionValue(const TimingData &Timing, float Beat);
+double SectionValue(const TimingData &Timing, double Beat);
 
 /*
 	Find the time at Beat given a constant function defined in intervals by Timing
 */
-double TimeAtBeat(const TimingData &Timing, float Offset, float Beat);
+double TimeAtBeat(const TimingData &Timing, float Offset, double Beat);
 
 /*
 	Find the antiderivative of an interval defined function through the Timing variable
 	from 0 to Time.
 */
-double IntegrateToTime(const TimingData &Timing, float Time, float Drift = 0);
+double IntegrateToTime(const TimingData &Timing, double Time, float Drift = 0);
 
 /*
 	Find the sum of all stops that have happened in [0, Beat)
 */
-double StopTimeAtBeat(const TimingData &StopsTiming, float Beat);
+double StopTimeAtBeat(const TimingData &StopsTiming, double Beat);
 
-double BeatAtTime(const TimingData &Timing, float Time, float Offset);
+double BeatAtTime(const TimingData &Timing, double Time, float Offset);
 
 #define DifficultyDuration(MySong, Diff) \
 	(Diff.Measures.size()) ? \
